backend/pyproject.toml
[project]
name = "admission-scraper-api"
version = "0.1.0"
description = "A RESTful API for accessing admission announcements and related programs from educational institutions."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.136.1",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
]
